+++ sdk/core/Azure.Core.Expressions.DataFactory/src/DataFactorySecretBaseDefinition.cs
@@ -8,16 +8,18 @@ namespace Azure.Core.Expressions.DataFactory
     /// Please note <see cref="DataFactorySecretBaseDefinition"/> is the base class. According to the scenario, a derived class of the base class might need to be assigned here, or this property needs to be casted to one of the possible derived classes.
     /// The available derived classes include <see cref="DataFactorySecretString"/> and <see cref="DataFactoryKeyVaultSecretReference"/>.
     /// </summary>
-    [TypeReferenceType(false, new[]{ nameof(SecretBaseType)})]
+    [PropertyReferenceType(new string[0], new[]{ nameof(SecretBaseType)})]
     public abstract partial class DataFactorySecretBaseDefinition
     {
         /// <summary> Initializes a new instance of DataFactorySecretBaseDefinition. </summary>
+        [InitializationConstructor]
         protected DataFactorySecretBaseDefinition()
         {
         }
 
         /// <summary> Initializes a new instance of DataFactorySecretBaseDefinition. </summary>
         /// <param name="secretBaseType"> Type of the secret. </param>
+        [SerializationConstructor]
         internal DataFactorySecretBaseDefinition(string? secretBaseType)
         {
             SecretBaseType = secretBaseType;