+++ sdk/storage/Azure.Storage.DataMovement/CHANGELOG.md
@@ -1,6 +1,16 @@
 # Release History
 
-## 12.0.0-beta.3 (Unreleased)
+## 12.0.0-beta.4 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
+## 12.0.0-beta.3 (2023-07-11)
 
 ### Features Added
 - `TransferManager` new API `PauseAllRunningTransfersAsync`.
@@ -8,6 +18,7 @@
 - Added support for tracking progress of transfers. See `TransferOptions.ProgressHandler` and `TransferOptions.ProgressHandlerOptions`.
 - Added `TransferManager.GetResumableTransfers` to get information about transfers that can be resumed.
 - Added support for `Transfermanager.ResumeTransferAsync` to resume a transfer.
+- Added support authorization using Azure Active Directory when using Service to Service Copy. 
 
 ### Breaking Changes
 - [BREAKING CHANGE] Altered API signatures on `TransferManager` and `DataTransfer` for pausing.
@@ -17,14 +28,13 @@
 - [BREAKING CHANGE] Combined both `TransferManager.StartTransferAsync` methods into one that accepts single or container resources. All existing calls should continue to work due to inheritence.
 - [BREAKING CHANGE] Renamed `ErrorHandlingOptions` to `ErrorHandlingBehavior`.
 - [BREAKING CHANGE] Changed type of `StorageResource.CanProduceUri` to `bool`.
+- [BREAKING CHANGE] Removed `TransferOptions.ResumeFromCheckpointId`. Use `Transfermanager.ResumeTransferAsync` to resume a transfer instead.
 
 ### Bugs Fixed
 - Fix to prevent empty strings or null to be passed as paths for `LocalFileStorageResource` and `LocalDirectoryStorageResourceContainer`.
 - Fixed `ErrorHandlingOptions.ContinueOnFailure` not be respected.
 - Fixed bug where resuming a transfer where the source and destination is a `StorageResourceContainer` would throw a null reference exception. 
 - Fixed bug when downloading zero length `StorageResource`s in a `StorageResourceContainer` will throw an exception.
-
-### Other Changes
 
 ## 12.0.0-beta.2 (2023-04-26)
 - [BREAKING CHANGE] Combined `SingleTransferOptions` and `ContainerTransferOptions` into `TransferOptions`.