+++ sdk/storage/Azure.Storage.Files.Shares/CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release History
 
-## 12.15.0-beta.2 (Unreleased)
+## 12.16.0-beta.1 (Unreleased)
 
 ### Features Added
 
@@ -9,6 +9,9 @@
 ### Bugs Fixed
 
 ### Other Changes
+
+## 12.15.0 (2023-07-23)
+- Includes all features from 12.15.0-beta.1.
 
 ## 12.15.0-beta.1 (2023-05-30)
 - Added support for service version 2023-01-03.